refactor(preferences): move migration modules to domain-specific structure
Move all preference-related migration modules from `services/configMigration/`
to `services/preferences/migrations/` and update corresponding imports in
tests, services, and e2e files to reflect the new structure.
